requirements.txt
ipywidgets
numpy
matplotlib
pyarrow
pandas
pyodbc
SQLalchemy
pyarrow
openpyxl
polars
seaborn
ydata-profiling
statsmodels
networkx[default]
tqdm
python-dotenv